requirements.txt
Flask>=2.0.0
python-dotenv>=0.19.0
Pillow>=9.0.0
# Optional but recommended for web apps:
Flask-Login>=0.5.0
Flask-WTF>=1.0.0
